infrastructureモジュールの整理 - annotationモジュール名をdetectorへ変更 - migrateモジュールをgeneratorとexecutorへ分離 - generatorとexecutorモジュールのトランザクションをSpringの仕組みに任せる
