--- backend/infrastructure/migration/annotation/src/main/kotlin/net/brightroom/migration/Migratable.kt
@@ -1,7 +1,0 @@
-package net.brightroom.migration
-
-@Target(AnnotationTarget.PROPERTY, AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Migratable(
-    val order: Int = 0,
-)
