--- backend/infrastructure/migration/annotation/src/test/kotlin/net/brightroom/migration/TableDiscoveryTest.kt
@@ -1,15 +1,0 @@
-package net.brightroom.migration
-
-import org.junit.jupiter.api.Test
-import kotlin.test.assertEquals
-
-class TableDiscoveryTest {
-    @Test
-    fun test() {
-        val tables = TableDiscovery.findAnnotatedTableObjects("net.brightroom.migration")
-
-        assertEquals(2, tables.size)
-        assertEquals("migration_test_1", tables[0].tableName)
-        assertEquals("migration_test_2", tables[1].tableName)
-    }
-}
