+++ backend/infrastructure/migration/detector/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+}
+
+dependencies {
+    implementation(libs.exposed.core)
+
+    testImplementation(libs.kotlin.test.junit5)
+}
+
+kotlin {
+    jvmToolchain(21)
+}
+
+tasks {
+    test {
+        useJUnitPlatform()
+    }
+}
