+++ backend/infrastructure/migration/detector/src/main/kotlin/net/brightroom/migration/detector/Migratable.kt
@@ -1,0 +1,5 @@
+package net.brightroom.migration.detector
+
+@Target(AnnotationTarget.PROPERTY, AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Migratable
