+++ backend/infrastructure/migration/detector/src/test/kotlin/net/brightroom/migration/detector/TableDiscoveryTest.kt
@@ -1,0 +1,15 @@
+package net.brightroom.migration.detector
+
+import org.junit.jupiter.api.Test
+import kotlin.test.assertEquals
+
+class TableDiscoveryTest {
+    @Test
+    fun test() {
+        val tables = TableDiscovery.findAnnotatedTableObjects("net.brightroom.migration.detector")
+
+        assertEquals(2, tables.size)
+        assertEquals("migration_test_1", tables[0].tableName)
+        assertEquals("migration_test_2", tables[1].tableName)
+    }
+}
