+++ backend/infrastructure/migration/executor/src/main/kotlin/net/brightroom/migration/executor/domain/model/MigrationPackage.kt
@@ -1,0 +1,12 @@
+package net.brightroom.migration.executor.domain.model
+
+/**
+ *
+ */
+class MigrationPackage(
+    private val value: String,
+) {
+    operator fun invoke(): String = value
+
+    override fun toString(): String = value
+}
