+++ backend/infrastructure/migration/executor/src/main/resources/application-default.yaml
@@ -1,0 +1,10 @@
+---
+spring:
+  datasource:
+    url: "jdbc:postgresql://localhost:15432/todo?currentSchema=public"
+    username: "brapl"
+    password: "brapl"
+    driver-class-name: "org.postgresql.Driver"
+logging:
+  level:
+    net.brightroom: "debug"
