+++ backend/infrastructure/migration/generator/src/main/kotlin/net/brightroom/migration/generator/_configuration/MigrationProperties.kt
@@ -1,0 +1,17 @@
+package net.brightroom.migration.generator._configuration
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.bind.ConstructorBinding
+
+@ConfigurationProperties("migration")
+data class MigrationProperties
+    @ConstructorBinding
+    constructor(
+        val migratePackage: String,
+        val generateScript: GenerateScriptProperties,
+    )
+
+data class GenerateScriptProperties(
+    val outputDirectory: String,
+    val filenameSuffix: String,
+)
