+++ backend/infrastructure/migration/generator/src/main/kotlin/net/brightroom/migration/generator/application/repository/CreateMigrationScriptRepository.kt
@@ -1,0 +1,7 @@
+package net.brightroom.migration.generator.application.repository
+
+import org.jetbrains.exposed.v1.core.Table
+
+interface CreateMigrationScriptRepository {
+    fun create(tables: Array<Table>)
+}
