+++ backend/infrastructure/migration/generator/src/main/kotlin/net/brightroom/migration/generator/infrastructure/MigrationScriptFileGenerator.kt
@@ -1,0 +1,27 @@
+package net.brightroom.migration.generator.infrastructure
+
+import net.brightroom.migration.generator.application.repository.CreateMigrationScriptRepository
+import net.brightroom.migration.generator.domain.model.OutputDirectory
+import net.brightroom.migration.generator.domain.model.ScriptName
+import org.jetbrains.exposed.v1.core.ExperimentalDatabaseMigrationApi
+import org.jetbrains.exposed.v1.core.Table
+import org.jetbrains.exposed.v1.migration.MigrationUtils
+import org.springframework.stereotype.Repository
+import org.springframework.transaction.annotation.Transactional
+
+@Repository
+class MigrationScriptFileGenerator(
+    private val outputDirectory: OutputDirectory,
+    private val scriptName: ScriptName,
+) : CreateMigrationScriptRepository {
+    @OptIn(ExperimentalDatabaseMigrationApi::class)
+    @Transactional(readOnly = true)
+    override fun create(tables: Array<Table>) {
+        MigrationUtils.generateMigrationScript(
+            *tables,
+            scriptDirectory = outputDirectory(),
+            scriptName = scriptName(),
+            withLogs = false,
+        )
+    }
+}
