+++ backend/infrastructure/migration/generator/src/main/resources/application.yaml
@@ -1,0 +1,17 @@
+---
+spring:
+  application:
+    name: "migration-generator"
+  datasource:
+    url: "jdbc:postgresql://localhost:15432/todo?currentSchema=public"
+    username: "brapl"
+    password: "brapl"
+    driver-class-name: "org.postgresql.Driver"
+logging:
+  level:
+    net.brightroom: "debug"
+migration:
+  migrate-package: "net.brightroom.schemas.schema.todo"
+  generate-script:
+    output-directory: "../executor/src/main/resources/migration"
+    filename-suffix: "generated_migration_script"
