--- backend/infrastructure/migration/migrate/src/main/resources/application-default.yaml
@@ -1,10 +1,0 @@
----
-logging:
-  level:
-    net.brightroom: "debug"
-migration:
-  datasource:
-    url: "jdbc:postgresql://localhost:15432/todo?currentSchema=public"
-    username: "brapl"
-    password: "brapl"
-    driver-class-name: "org.postgresql.Driver"
