--- backend/infrastructure/migration/migrate/src/main/resources/application.yaml
@@ -1,18 +1,0 @@
----
-spring:
-  application:
-    name: "database-migration"
-logging:
-  level:
-    Exposed: "debug"
-    net.brightroom: "info"
-migration:
-  datasource:
-    url: "${DATASOURCE_TODO_URL}"
-    username: "${DATASOURCE_TODO_USERNAME}"
-    password: "${DATASOURCE_TODO_PASSWORD}"
-    driver-class-name: "org.postgresql.Driver"
-  migrate-package: "net.brightroom.schemas.schema.todo"
-  generate-script:
-    output-directory: "src/main/resources/migration"
-    filename-suffix: "generated_migration_script"
