+++ backend/infrastructure/schemas/src/main/kotlin/net/brightroom/schemas/schema/todo/PasswordTable.kt
@@ -2,13 +2,13 @@
 
 package net.brightroom.schemas.schema.todo
 
-import net.brightroom.migration.Migratable
+import net.brightroom.migration.detector.Migratable
 import org.jetbrains.exposed.v1.core.dao.id.IntIdTable
 import org.jetbrains.exposed.v1.datetime.CurrentDateTime
 import org.jetbrains.exposed.v1.datetime.datetime
 import kotlin.uuid.ExperimentalUuidApi
 
-@Migratable(order = 1)
+@Migratable
 object PasswordTable : IntIdTable("password") {
     val account_id = reference("account_id", AccountIdTable)
     val hash = varchar("hash", 255)
