+++ backend/infrastructure/schemas/src/main/kotlin/net/brightroom/schemas/schema/todo/TaskContentTable.kt
@@ -2,13 +2,13 @@
 
 package net.brightroom.schemas.schema.todo
 
-import net.brightroom.migration.Migratable
+import net.brightroom.migration.detector.Migratable
 import org.jetbrains.exposed.v1.core.dao.id.IntIdTable
 import org.jetbrains.exposed.v1.datetime.CurrentDateTime
 import org.jetbrains.exposed.v1.datetime.datetime
 import kotlin.uuid.ExperimentalUuidApi
 
-@Migratable(order = 2)
+@Migratable
 object TaskContentTable : IntIdTable("task_content") {
     val task_id = reference("task_id", TaskIdTable)
     val title = varchar("title", 200)
