+++ backend/infrastructure/schemas/src/main/kotlin/net/brightroom/schemas/schema/todo/TaskIdTable.kt
@@ -2,13 +2,13 @@
 
 package net.brightroom.schemas.schema.todo
 
-import net.brightroom.migration.Migratable
+import net.brightroom.migration.detector.Migratable
 import net.brightroom.schemas._extensions.exposed.KotlinUUIDTable
 import org.jetbrains.exposed.v1.datetime.CurrentDateTime
 import org.jetbrains.exposed.v1.datetime.datetime
 import kotlin.uuid.ExperimentalUuidApi
 
-@Migratable(order = 1)
+@Migratable
 object TaskIdTable : KotlinUUIDTable("task_id") {
     val account_id = reference("account_id", AccountIdTable)
     val createdAt = datetime("created_at").defaultExpression(CurrentDateTime).index()
