+++ backend/infrastructure/schemas/src/main/kotlin/net/brightroom/schemas/schema/todo/TaskStatusTable.kt
@@ -2,14 +2,14 @@
 
 package net.brightroom.schemas.schema.todo
 
-import net.brightroom.migration.Migratable
+import net.brightroom.migration.detector.Migratable
 import net.brightroom.todo.domain.model.task.status.Status
 import org.jetbrains.exposed.v1.core.dao.id.IntIdTable
 import org.jetbrains.exposed.v1.datetime.CurrentDateTime
 import org.jetbrains.exposed.v1.datetime.datetime
 import kotlin.uuid.ExperimentalUuidApi
 
-@Migratable(order = 2)
+@Migratable
 object TaskStatusTable : IntIdTable("task_status") {
     val taskId = reference("task_id", TaskIdTable)
     val status = enumerationByName<Status>("status", 5)
