+++ backend/infrastructure/schemas/src/main/kotlin/net/brightroom/schemas/transaction/Transaction.kt
@@ -1,19 +1,28 @@
 package net.brightroom.schemas.transaction
 
+import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.runBlocking
 import org.jetbrains.exposed.v1.core.StdOutSqlLogger
 import org.jetbrains.exposed.v1.r2dbc.R2dbcDatabase
 import org.jetbrains.exposed.v1.r2dbc.transactions.suspendTransaction
+import org.jetbrains.exposed.v1.r2dbc.transactions.suspendTransactionAsync
 
-inline fun <T> transaction(
+suspend inline fun <T> transaction(
     db: R2dbcDatabase,
-    readOnly: Boolean = true,
+    readOnly: Boolean = false,
     crossinline block: suspend () -> T,
 ): T =
-    runBlocking {
-        suspendTransaction(context = Dispatchers.IO, readOnly = readOnly, db = db) {
-            addLogger(StdOutSqlLogger)
-            block()
-        }
+    suspendTransaction(context = Dispatchers.IO, readOnly = readOnly, db = db) {
+        addLogger(StdOutSqlLogger)
+        block()
+    }
+
+suspend inline fun <T> transactionAsync(
+    db: R2dbcDatabase,
+    readOnly: Boolean = false,
+    crossinline block: suspend () -> T,
+): Deferred<T> =
+    suspendTransactionAsync(context = Dispatchers.IO, readOnly = readOnly, db = db) {
+        addLogger(StdOutSqlLogger)
+        block()
     }
