+++ settings.gradle.kts
@@ -46,8 +46,9 @@ include(":shared")
 include(":backend:domain")
 
 include(":backend:infrastructure:schemas")
-include(":backend:infrastructure:migration:annotation")
-include(":backend:infrastructure:migration:migrate")
+include(":backend:infrastructure:migration:detector")
+include(":backend:infrastructure:migration:generator")
+include(":backend:infrastructure:migration:executor")
 
 include(":backend:application:api")
 include(":backend:application:scheduler")
